third_party/blink/web_tests/flag-specific/highdpi/animations/interpolation/border-image-slice-interpolation-stability-expected.txt

This is a testharness.js-based test.
FAIL border-image-slice-interpolation-stability assert_equals: expected "50" but got "75"
Harness: the test ran to completion.
